+++ src/com/ms/silverking/net/async/SimpleNewConnectionTimeoutController.java
@@ -60,12 +60,12 @@ public class SimpleNewConnectionTimeoutController implements NewConnectionTimeou
     }
 
     @Override
-    public int getRelativeTimeoutMillisForAttempt(AddrAndPort addrAndPort, int attemptIndex) {
+    public long getRelativeTimeoutMillisForAttempt(AddrAndPort addrAndPort, int attemptIndex) {
         return attemptRelativeTimeoutMillis;
     }
 
     @Override
-    public int getMaxRelativeTimeoutMillis(AddrAndPort addrAndPort) {
+    public long getMaxRelativeTimeoutMillis(AddrAndPort addrAndPort) {
         return maxRelativeTimeoutMillis;
     }
